Instruction simplification is no lognger an Instruction method
