+++ compiler/noirc_evaluator/src/ssa/ir/dfg.rs
@@ -2,7 +2,6 @@ use std::{borrow::Cow, sync::Arc};
 
 use crate::ssa::{
     function_builder::data_bus::DataBus,
-    ir::instruction::SimplifyResult,
     opt::pure::{FunctionPurities, Purity},
 };
 
@@ -24,6 +23,9 @@ use iter_extended::vecmap;
 use serde::{Deserialize, Serialize};
 use serde_with::DisplayFromStr;
 use serde_with::serde_as;
+use simplify::{SimplifyResult, simplify};
+
+mod simplify;
 
 /// The DataFlowGraph contains most of the actual data in a function including
 /// its blocks, instructions, and values. This struct is largely responsible for
@@ -314,7 +316,7 @@ impl DataFlowGraph {
             return InsertInstructionResult::InstructionRemoved;
         }
 
-        match instruction.simplify(self, block, ctrl_typevars.clone(), call_stack) {
+        match simplify(&instruction, self, block, ctrl_typevars.clone(), call_stack) {
             SimplifyResult::SimplifiedTo(simplification) => {
                 InsertInstructionResult::SimplifiedTo(simplification)
             }
