+++ compiler/noirc_evaluator/src/ssa/ir/dfg/simplify/constrain.rs
@@ -1,8 +1,11 @@
 use acvm::{FieldElement, acir::AcirField};
 
-use crate::ssa::ir::types::NumericType;
-
-use super::{Binary, BinaryOp, ConstrainError, DataFlowGraph, Instruction, Type, Value, ValueId};
+use crate::ssa::ir::{
+    dfg::DataFlowGraph,
+    instruction::{Binary, BinaryOp, ConstrainError, Instruction},
+    types::{NumericType, Type},
+    value::{Value, ValueId},
+};
 
 /// Try to decompose this constrain instruction. This constraint will be broken down such that it instead constrains
 /// all the values which are used to compute the values which were being constrained.
@@ -224,6 +227,38 @@ mod tests {
             return v0
         }
         ";
+        assert_normalized_ssa_equals(ssa, expected);
+    }
+
+    #[test]
+    fn constraint_decomposition() {
+        // When constructing this IR, we should automatically decompose the constraint to be in terms of `v0`, `v1` and `v2`.
+        //
+        // The mul instructions are retained and will be removed in the dead instruction elimination pass.
+        let src = "
+            acir(inline) fn main f0 {
+              b0(v0: u1, v1: u1, v2: u1):
+                v3 = mul v0, v1
+                v4 = not v2
+                v5 = mul v3, v4
+                constrain v5 == u1 1
+                return
+            }
+            ";
+        let ssa = Ssa::from_str_simplifying(src).unwrap();
+
+        let expected = "
+            acir(inline) fn main f0 {
+              b0(v0: u1, v1: u1, v2: u1):
+                v3 = mul v0, v1
+                v4 = not v2
+                v5 = mul v3, v4
+                constrain v0 == u1 1
+                constrain v1 == u1 1
+                constrain v2 == u1 0
+                return
+            }
+            ";
         assert_normalized_ssa_equals(ssa, expected);
     }
 }
